refactor(client/kotlin): remove interface layer that seemed redundant
